% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiBitR.R
\docType{package}
\name{BiBitR}
\alias{BiBitR}
\alias{BiBitR-package}
\title{A biclustering algorithm for extracting bit-patterns from binary datasets}
\description{
BiBit R is a simple R wrapper which directly calls the original Java code for applying the BiBit algorithm.
The original Java code can be found at \url{http://eps.upo.es/bigs/BiBit.html} by Domingo S. Rodriguez-Baena, Antonia J. Perez-Pulido and Jesus S. Aguilar-Ruiz.
}
\references{
Domingo S. Rodriguez-Baena, Antonia J. Perez-Pulido and Jesus S. Aguilar-Ruiz (2011), "A biclustering algorithm for extracting bit-patterns from binary datasets", \emph{Bioinformatics}
}

