% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowfunction.R
\name{ExactFisherBC}
\alias{ExactFisherBC}
\title{Apply Fisher Exact Test on Biclusters of a Biclust object}
\usage{
ExactFisherBC(result, matrix, p.adjust = "BH", alpha = 0.05,
  BC = 1:result@Number)
}
\arguments{
\item{result}{A Biclust Object.}

\item{matrix}{Accompanying binary data matrix which was used to obtain \code{result}.}

\item{p.adjust}{Which method to use when adjusting p-values, see \code{\link[stats]{p.adjust}} (default=\code{"BH"}).}

\item{alpha}{Significance level (default=0.05).}

\item{BC}{Numeric vector to select for which BC's the Fisher Exact Test needs to be computed. Default is all available biclusters.}
}
\value{
Returns a list with two elements:
\itemize{
\item \code{summary}: a data frame containing the number of rows, significant rows, adjusted significant rows, columns, significant columns and adjusted significant columns for all requested biclusters.
\item \code{info}: a list with an element for each requested biclusters. Each BC list element contains two data frames (\code{row} and \code{col}) which contain the index, name, pvalue, adjusted pvalue, density of 1's inside and density of 1's outside for all the row and column members of the bicluster.
}
}
\description{
Accepts a Biclust Object and computes the Fisher Exact Test of the rows and columns inside the biclusters versus the rows and columns outside. 
This test gives some information on the fact if the rows or columns are uniquely active for this particular (or other similar) bicluster.
The function will \strong{not} extract the column pattern and test every row of the dataset. This functionality can be found in \code{\link{RowTest_Fisher}}.
}
\examples{
\dontrun{
data <- matrix(sample(c(0,1),100*100,replace=TRUE,prob=c(0.9,0.1)),nrow=100,ncol=100)
data[1:10,1:10] <- 1 # BC1
data[11:20,11:20] <- 1 # BC2
data[21:30,21:30] <- 1 # BC3
data <- data[sample(1:nrow(data),nrow(data)),sample(1:ncol(data),ncol(data))]

result1 <- bibit2(data,minr=5,minc=5,noise=0.1)
out_fisher <- ExactFisherBC(result1,data)
out_fisher$summary
out_fisher$info[[1]]
}
 
}
\author{
Ewoud De Troyer
}
