% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowfunction.R
\name{NoiseInfoBC}
\alias{NoiseInfoBC}
\title{Noise Info for Biclusters}
\usage{
NoiseInfoBC(result, matrix, plot = FALSE, plot.BC = 1:result@Number)
}
\arguments{
\item{result}{A Biclust Object.}

\item{matrix}{Accompanying binary data matrix which was used to obtain \code{result}.}

\item{plot}{Boolean value (default=FALSE) to create bar plots of the number of rows which have 0, 1, 2,... noise.}

\item{plot.BC}{Vector for which BC's the barplots need to be created. (default = all biclusters)}
}
\value{
A data frame containing the following variables for all BC's: Row/Column Dimension, Maximum Row Noise and how many of the rows fit with 0 noise, 1 noise,...
}
\description{
Collect some info on the row noise distribution of each Bicluster of a Biclust or BiBitWorkflow (\code{$Biclust} slot) object. 
The information collected are the row and column dimension, the maximum row noise and the number of rows which 0, 1, 2,... noise.
}
\examples{
\dontrun{
data <- matrix(sample(c(0,1),100*100,replace=TRUE,prob=c(0.9,0.1)),nrow=100,ncol=100)
data[1:10,1:10] <- 1 # BC1
data[11:20,11:20] <- 1 # BC2
data[21:30,21:30] <- 1 # BC3
data <- data[sample(1:nrow(data),nrow(data)),sample(1:ncol(data),ncol(data))]
result <- bibit2(data,minr=5,minc=5,noise=1)
NoiseInfoBC(result=result,matrix=data)
}
}
\author{
Ewoud De Troyer
}
