% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{methods}
\name{summary,Biclust-method}
\alias{summary,Biclust-method}
\title{Summary Method for Biclust Class}
\usage{
\S4method{summary}{Biclust}(object)
}
\arguments{
\item{object}{Biclust S4 Object}
}
\description{
Summary Method for Biclust Class
}

