% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bibit3_patternBC}
\alias{bibit3_patternBC}
\title{Extract BC from \code{bibit3} result and add pattern}
\usage{
bibit3_patternBC(result, matrix, pattern = c(1), type = c("full", "sub",
  "ext"), BC = c(1))
}
\arguments{
\item{result}{Result produced by \code{\link{bibit3}}}

\item{matrix}{The binary input matrix.}

\item{pattern}{Vector containing either the number or name of which patterns the BC results should be extracted.}

\item{type}{Vector for which BC results should be printed.
\itemize{
\item Full Pattern (\code{"full"})
\item Sub Pattern (\code{"sub"})
\item Extended (\code{"ext"})
}}

\item{BC}{Vector of BC indices which should be printed, conditioned on \code{pattern} and \code{type}.}
}
\value{
Prints querried biclusters.
}
\description{
Function which will print the BC matrix and add 2 duplicate articial pattern rows on top. The function allows you to see the BC and the pattern the BC was guided towards to.
}
\examples{
\dontrun{
# TO DO
}
}
\author{
Ewoud De Troyer
}

