% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowfunction.R
\name{NoiseScree}
\alias{NoiseScree}
\title{Noise Scree Plots}
\usage{
NoiseScree(result, matrix, type = c("Added", "Total"), pattern = NULL,
  noise_select = 0, alpha = 0.05)
}
\arguments{
\item{result}{A Biclust or BiBitWorkflow Object.}

\item{matrix}{Accompanying binary data matrix which was used to obtain \code{result}.}

\item{type}{Either \code{"Added"} or \code{"Total"}. Should the noise level be plotted against the number of added rows (at that noise level) or the total number of rows (up to that noise level)?}

\item{pattern}{Numeric vector for which patterns the noise scree plot should be drawn (default = all patterns).}

\item{noise_select}{Should an automatic noise selection be applied and drawn (blue vertical line) on the plot? (Using ad hoc method to find the elbow/kink in the Noise Scree plots)
\itemize{
\item \code{noise_select=0}: No noise selection is applied and no line is drawn (default).
\item \code{noise_select=1}: Using the Noise Scree plot (with 'Added Rows' on the y-axis), find the noise level where the current number of added rows at this noise level is larger than the mean of 'added rows' at the lower noise levels. 
After locating this noise level, lower the noise level by 1. This is your automatically selected elbow/kink and therefore your noise level.
\item \code{noise_select=2}: Applies the same steps as for \code{noise_select=1}, but instead of decreasing the noise level by only 1, keep decreasing the noise level until the number of added rows isn't decreasing anymore either.
}}

\item{alpha}{If info from the Fisher Exact test is available, which significance level should be used to in the plot (Noise versus Significant Fisher Exact Test rows). (default=0.05)}
}
\value{
\code{NULL}
}
\description{
Extract patterns from either a Biclust or BiBitWorkflow object (see Details) and plot the Noise Scree plot (same as plot 4 in \code{\link{BiBitWorkflow}}). Additionally, if \code{FisherResult} is available (from \code{\link{RowTest_Fisher}}), this info will be added to the plot.
}
\details{
\describe{
\item{\emph{Biclust S4 Object}}{

Using the column patterns of the Biclust result, the noise level is plotted versus the number of \code{"Total"} or \code{"Added"} rows.
}
\item{\emph{BiBitWorkflow S3 Object}}{

The merged column patterns (after cutting the hierarchical tree) are extracted from the BiBitWorkflow object, namely the \code{$info$MergedColPatterns} slot. 
These patterns are used to plot the noise level versus the number of \code{"Total"} or \code{"Added"} rows.
}
}
If information on the Fisher Exact Test is available, then this info will added to the plot (noise level versus significant rows).
}
\examples{
\dontrun{
test
}
}
\author{
Ewoud De Troyer
}

